Towards wait notify implementation
Support addition and removal to and from the waiters queue of a monitor
in the MMGR.

Add functionality to release a monitor and be added to the waiters queue
with a single message.

In order not to implement yet another malloc and not to through away
space we use the same allocator as for the monitors and we pack two
waiters per allocation chunk.