Avoid exiting when there are no more threads.
Each core holds its own isolate.  As a result, it might detect that
there are no more threads to execute in the isolate.  In Squawk this
signals the termination of the application.  However, in our case we
might just be idle and need to wait for new jobs.  FIXME: Implement
proper finalization of applications.