Remove ClassOrAssociation
ClassOrAssociation used to be an intermediate object holding the
metadata of an object, including the monitor.  This object was only
created when needed.  Instead we add a monitor object in the Object
class and allocate it only when needed.  This way we reduce the
complexity of the code and the pointer chaining (which resulted in more
fetches)